+++ src/client/kotlin/net/easecation/bedrockloader/render/model/ModelCuboidData.kt
@@ -5,16 +5,16 @@ package net.easecation.bedrockloader.render.model
 
 import net.fabricmc.api.EnvType
 import net.fabricmc.api.Environment
-import net.minecraft.client.util.math.Vector2f
-import org.joml.Vector3f
+import org.joml.Vector2d
+import org.joml.Vector3d
 
 @Environment(value = EnvType.CLIENT)
-class ModelCuboidData(private val name: String?, textureX: Float, textureY: Float, private val faceUV: ModelPart.FaceUV?, offsetX: Float, offsetY: Float, offsetZ: Float, sizeX: Float, sizeY: Float, sizeZ: Float, extra: Dilation, private val mirror: Boolean, textureScaleX: Float, textureScaleY: Float) {
-    private val offset = Vector3f(offsetX, offsetY, offsetZ)
-    private val dimensions = Vector3f(sizeX, sizeY, sizeZ)
+class ModelCuboidData(private val name: String?, textureX: Double, textureY: Double, private val faceUV: ModelPart.FaceUV?, offsetX: Double, offsetY: Double, offsetZ: Double, sizeX: Double, sizeY: Double, sizeZ: Double, extra: Dilation, private val mirror: Boolean, textureScaleX: Double, textureScaleY: Double) {
+    private val offset = Vector3d(offsetX, offsetY, offsetZ)
+    private val dimensions = Vector3d(sizeX, sizeY, sizeZ)
     private val extraSize = extra
-    private val textureUV = Vector2f(textureX, textureY)
-    private val textureScale = Vector2f(textureScaleX, textureScaleY)
+    private val textureUV = Vector2d(textureX, textureY)
+    private val textureScale = Vector2d(textureScaleX, textureScaleY)
 
     fun createCuboid(textureWidth: Int, textureHeight: Int): ModelPart.Cuboid {
         return ModelPart.Cuboid(
@@ -31,8 +31,8 @@ class ModelCuboidData(private val name: String?, textureX: Float, textureY: Floa
             extraSize.radiusY,
             extraSize.radiusZ,
             this.mirror,
-            textureWidth.toFloat() * textureScale.x,
-            textureHeight.toFloat() * textureScale.y
+            textureWidth.toDouble() * textureScale.x,
+            textureHeight.toDouble() * textureScale.y
         )
     }
 }
