+++ src/client/kotlin/net/easecation/bedrockloader/render/model/ModelTransform.kt
@@ -7,18 +7,18 @@ import net.fabricmc.api.EnvType
 import net.fabricmc.api.Environment
 
 @Environment(value = EnvType.CLIENT)
-class ModelTransform private constructor(val pivotX: Float, val pivotY: Float, val pivotZ: Float, val pitch: Float, val yaw: Float, val roll: Float, val detachPivot: Boolean = false) {
+class ModelTransform private constructor(val pivotX: Double, val pivotY: Double, val pivotZ: Double, val pitch: Double, val yaw: Double, val roll: Double, val detachPivot: Boolean = false) {
     companion object {
-        val NONE: ModelTransform = of(0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f)
-        fun pivot(pivotX: Float, pivotY: Float, pivotZ: Float): ModelTransform {
-            return of(pivotX, pivotY, pivotZ, 0.0f, 0.0f, 0.0f)
+        val NONE: ModelTransform = of(0.0, 0.0, 0.0, 0.0, 0.0, 0.0)
+        fun pivot(pivotX: Double, pivotY: Double, pivotZ: Double): ModelTransform {
+            return of(pivotX, pivotY, pivotZ, 0.0, 0.0, 0.0)
         }
 
-        fun rotation(pitch: Float, yaw: Float, roll: Float): ModelTransform {
-            return of(0.0f, 0.0f, 0.0f, pitch, yaw, roll)
+        fun rotation(pitch: Double, yaw: Double, roll: Double): ModelTransform {
+            return of(0.0, 0.0, 0.0, pitch, yaw, roll)
         }
 
-        fun of(pivotX: Float, pivotY: Float, pivotZ: Float, pitch: Float, yaw: Float, roll: Float, detachPivot: Boolean = false): ModelTransform {
+        fun of(pivotX: Double, pivotY: Double, pivotZ: Double, pitch: Double, yaw: Double, roll: Double, detachPivot: Boolean = false): ModelTransform {
             return ModelTransform(pivotX, pivotY, pivotZ, pitch, yaw, roll, detachPivot)
         }
     }
